Waiting turns, Stockfish engine
Players may no longer interact with the opponent's pieces.

Saved scripts to the repository for the functionality of the auto-calculating chess move engine.

Instantiated an engine object with a method for reading the board positioning to calculate the best move, and an event handler for transforming the notation data into variables usable in the pre-existing game.move() logic.