Game over closure, new game
Implemented a pop-up modal to inform the user that the game has ended. The user then has the chance to start a new game, which resets the board and the engine's memory.
New css file created for styling the pop-up.

Users may also choose to reset the game using a menu button.

Restyled comment headers for clarity.