Controls - game state
The player may not move pieces until a game has started.
The player may not change menu controls while a game is in progress.

Needing to introduce a new modal escape button so the user may change the match settings instead of just rematching - starting a new game disables the controls so at present the user cannot change the difficulty after losing.