Click-moving, move highlights
Play may not click 2 squares to move: starting square (the piece they want to move), and target square (where they'd like to move it to.
The same legal-move checking for piece dragging is in place.

When a player clicks a piece (1st click), the legal squares that the piece may target are highlighted.