Eval bar & resign button adjustments
Added an evaluation score (measured in pawns) to the evaluation bar, giving meaning to the visualization.

Enabled on/off toggling for the evaluation score bar from within the in-game options menu.
The resign button has been relocated from the in-game options menu to the bottom control row of the game container, as to only include in-game accessibility options in the in-game options menu.