Piece animation, small fixes
Pieces of the players color now wiggle when hovering over them and after having moved squares.

Fixed a bug where the undo button was enabled after game completion.

Fixed a bug where the previous game's moves could be navigated through when rematching.

Adjusted the evaluation score bar to be flipped if the player is playing as Black.