Hint (best move)
The player may now use a button or hotkey (H) to request a hint, which highlights the best move for the player as calculated by the engine.

The hints provided have mostly been sound with some occasions of blundering as a result of the difficulty (adding dirt to the calculation filters). May need to configure a new engine at maximum difficulty for the purposes of receiving only the best hints.
Also to note that the engine hint calculation is not nearly as quick as the move execution on the engines turn. To investigate performance enhancement at a later date.